requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.36
jinja2>=3.1.4
python-multipart>=0.0.18
requests>=2.32.0
beautifulsoup4>=4.12.3
pydantic>=2.10.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
